@@ -1,6 +1,6 @@
 {
   "name": "crossfilter2",
-  "version": "1.4.3",
+  "version": "1.5.4",
   "description": "Fast multidimensional filtering for coordinated views.",
   "license": "Apache-2.0",
   "keywords": [
@@ -32,34 +32,44 @@
       "url": "https://github.com/esjewett"
     }
   ],
-  "homepage": "http://crossfilter.github.com/crossfilter/",
-  "main": "./index.js",
+  "homepage": "https://crossfilter.github.io/crossfilter/",
+  "main": "./crossfilter.js",
+  "module": "main.js",
+  "types": "./index.d.ts",
+  "unpkg": "./crossfilter.min.js",
   "repository": {
     "type": "git",
     "url": "http://github.com/crossfilter/crossfilter.git"
   },
   "dependencies": {
-    "lodash.result": "^4.4.0"
+    "@ranfdev/deepobj": "1.0.2"
   },
   "devDependencies": {
-    "browserify": "^13.0.0",
-    "d3": "3.5",
-    "eslint": "2.10.2",
+    "d3": "^3.5.17",
+    "eslint": "^6.5.1",
+    "esm": "^3.2.25",
     "package-json-versionify": "1.0.2",
-    "semver": "^5.3.0",
-    "sinon": "^4.0.2",
-    "uglify-js": "2.4.0",
-    "vows": "0.7.0"
+    "rollup": "1",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-json": "4",
+    "rollup-plugin-node-resolve": "5",
+    "rollup-plugin-terser": "5",
+    "semver": "^5.7.0",
+    "sinon": "^7.5.0",
+    "uglify-js": "^3.6.7",
+    "vows": "0.8.2"
   },
   "scripts": {
-    "benchmark": "node test/benchmark.js",
-    "build": "browserify index.js -t package-json-versionify --standalone crossfilter -o crossfilter.js && uglifyjs --compress --mangle --screw-ie8 crossfilter.js -o crossfilter.min.js",
+    "benchmark": "node -r esm test/benchmark.js",
+    "build": "rollup -c rollup.config.js",
     "clean": "rm -f crossfilter.js crossfilter.min.js",
-    "test": "./node_modules/.bin/vows --verbose && eslint src/"
+    "test": "node -r esm  ./node_modules/.bin/vows --verbose && eslint src/"
   },
   "files": [
-    "src",
     "index.js",
+    "main.js",
+    "src/**",
+    "index.d.ts",
     "crossfilter.js",
     "crossfilter.min.js"
   ],
