@@ -1,20 +1,25 @@
+let array8 = arrayUntyped,
+    array16 = arrayUntyped,
+    array32 = arrayUntyped,
+    arrayLengthen = arrayLengthenUntyped,
+    arrayWiden = arrayWidenUntyped;
 if (typeof Uint8Array !== "undefined") {
-  var crossfilter_array8 = function(n) { return new Uint8Array(n); };
-  var crossfilter_array16 = function(n) { return new Uint16Array(n); };
-  var crossfilter_array32 = function(n) { return new Uint32Array(n); };
+  array8 = function(n) { return new Uint8Array(n); };
+  array16 = function(n) { return new Uint16Array(n); };
+  array32 = function(n) { return new Uint32Array(n); };
 
-  var crossfilter_arrayLengthen = function(array, length) {
+  arrayLengthen = function(array, length) {
     if (array.length >= length) return array;
     var copy = new array.constructor(length);
     copy.set(array);
     return copy;
   };
 
-  var crossfilter_arrayWiden = function(array, width) {
+  arrayWiden = function(array, width) {
     var copy;
     switch (width) {
-      case 16: copy = crossfilter_array16(array.length); break;
-      case 32: copy = crossfilter_array32(array.length); break;
+      case 16: copy = array16(array.length); break;
+      case 32: copy = array32(array.length); break;
       default: throw new Error("invalid array width!");
     }
     copy.set(array);
@@ -22,25 +27,25 @@ if (typeof Uint8Array !== "undefined") {
   };
 }
 
-function crossfilter_arrayUntyped(n) {
+function arrayUntyped(n) {
   var array = new Array(n), i = -1;
   while (++i < n) array[i] = 0;
   return array;
 }
 
-function crossfilter_arrayLengthenUntyped(array, length) {
+function arrayLengthenUntyped(array, length) {
   var n = array.length;
   while (n < length) array[n++] = 0;
   return array;
 }
 
-function crossfilter_arrayWidenUntyped(array, width) {
+function arrayWidenUntyped(array, width) {
   if (width > 32) throw new Error("invalid array width!");
   return array;
 }
 
 // An arbitrarily-wide array of bitmasks
-function crossfilter_bitarray(n) {
+function bitarray(n) {
   this.length = n;
   this.subarrays = 1;
   this.width = 8;
@@ -48,25 +53,26 @@ function crossfilter_bitarray(n) {
     0: 0
   }
 
-  this[0] = crossfilter_array8(n);
+  this[0] = array8(n);
 }
 
-crossfilter_bitarray.prototype.lengthen = function(n) {
+bitarray.prototype.lengthen = function(n) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
-    this[i] = crossfilter_arrayLengthen(this[i], n);
+    this[i] = arrayLengthen(this[i], n);
   }
   this.length = n;
 };
 
 // Reserve a new bit index in the array, returns {offset, one}
-crossfilter_bitarray.prototype.add = function() {
+bitarray.prototype.add = function() {
   var m, w, one, i, len;
 
   for (i = 0, len = this.subarrays; i < len; ++i) {
     m = this.masks[i];
     w = this.width - (32 * i);
-    one = ~m & -~m;
+    // isolate the rightmost zero bit and return it as an unsigned int of 32 bits, if NaN or -1, return a 0 
+    one = (~m & (m + 1)) >>> 0;
 
     if (w >= 32 && !one) {
       continue;
@@ -74,7 +80,7 @@ crossfilter_bitarray.prototype.add = function() {
 
     if (w < 32 && (one & (1 << w))) {
       // widen this subarray
-      this[i] = crossfilter_arrayWiden(this[i], w <<= 1);
+      this[i] = arrayWiden(this[i], w <<= 1);
       this.width = 32 * i + w;
     }
 
@@ -87,7 +93,7 @@ crossfilter_bitarray.prototype.add = function() {
   }
 
   // add a new subarray
-  this[this.subarrays] = crossfilter_array8(this.length);
+  this[this.subarrays] = array8(this.length);
   this.masks[this.subarrays] = 1;
   this.width += 8;
   return {
@@ -97,7 +103,7 @@ crossfilter_bitarray.prototype.add = function() {
 };
 
 // Copy record from index src to index dest
-crossfilter_bitarray.prototype.copy = function(dest, src) {
+bitarray.prototype.copy = function(dest, src) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     this[i][dest] = this[i][src];
@@ -105,19 +111,18 @@ crossfilter_bitarray.prototype.copy = function(dest, src) {
 };
 
 // Truncate the array to the given length
-crossfilter_bitarray.prototype.truncate = function(n) {
+bitarray.prototype.truncate = function(n) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     for (var j = this.length - 1; j >= n; j--) {
       this[i][j] = 0;
     }
-    this[i].length = n;
   }
   this.length = n;
 };
 
 // Checks that all bits for the given index are 0
-crossfilter_bitarray.prototype.zero = function(n) {
+bitarray.prototype.zero = function(n) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     if (this[i][n]) {
@@ -128,7 +133,7 @@ crossfilter_bitarray.prototype.zero = function(n) {
 };
 
 // Checks that all bits for the given index are 0 except for possibly one
-crossfilter_bitarray.prototype.zeroExcept = function(n, offset, zero) {
+bitarray.prototype.zeroExcept = function(n, offset, zero) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     if (i === offset ? this[i][n] & zero : this[i][n]) {
@@ -138,9 +143,9 @@ crossfilter_bitarray.prototype.zeroExcept = function(n, offset, zero) {
   return true;
 };
 
-// Checks that all bits for the given indez are 0 except for the specified mask.
+// Checks that all bits for the given index are 0 except for the specified mask.
 // The mask should be an array of the same size as the filter subarrays width.
-crossfilter_bitarray.prototype.zeroExceptMask = function(n, mask) {
+bitarray.prototype.zeroExceptMask = function(n, mask) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     if (this[i][n] & mask[i]) {
@@ -151,7 +156,7 @@ crossfilter_bitarray.prototype.zeroExceptMask = function(n, mask) {
 }
 
 // Checks that only the specified bit is set for the given index
-crossfilter_bitarray.prototype.only = function(n, offset, one) {
+bitarray.prototype.only = function(n, offset, one) {
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     if (this[i][n] != (i === offset ? one : 0)) {
@@ -162,13 +167,13 @@ crossfilter_bitarray.prototype.only = function(n, offset, one) {
 };
 
 // Checks that only the specified bit is set for the given index except for possibly one other
-crossfilter_bitarray.prototype.onlyExcept = function(n, offset, zero, onlyOffset, onlyOne) {
+bitarray.prototype.onlyExcept = function(n, offset, zero, onlyOffset, onlyOne) {
   var mask;
   var i, len;
   for (i = 0, len = this.subarrays; i < len; ++i) {
     mask = this[i][n];
     if (i === offset)
-      mask &= zero;
+      mask = (mask & zero) >>> 0;
     if (mask != (i === onlyOffset ? onlyOne : 0)) {
       return false;
     }
@@ -176,11 +181,11 @@ crossfilter_bitarray.prototype.onlyExcept = function(n, offset, zero, onlyOffset
   return true;
 };
 
-module.exports = {
-  array8: crossfilter_arrayUntyped,
-  array16: crossfilter_arrayUntyped,
-  array32: crossfilter_arrayUntyped,
-  arrayLengthen: crossfilter_arrayLengthenUntyped,
-  arrayWiden: crossfilter_arrayWidenUntyped,
-  bitarray: crossfilter_bitarray
+export default {
+  array8: arrayUntyped,
+  array16: arrayUntyped,
+  array32: arrayUntyped,
+  arrayLengthen: arrayLengthenUntyped,
+  arrayWiden: arrayWidenUntyped,
+  bitarray: bitarray
 };
