@@ -1,7 +1,3 @@
-'use strict';
-
-function crossfilter_identity(d) {
+export default d => {
   return d;
-}
-
-module.exports = crossfilter_identity;
+};
