@@ -1,10 +1,6 @@
-'use strict';
-
-function permute(array, index, deep) {
+export default (array, index, deep) => {
   for (var i = 0, n = index.length, copy = deep ? JSON.parse(JSON.stringify(array)) : new Array(n); i < n; ++i) {
     copy[i] = array[index[i]];
   }
   return copy;
 }
-
-module.exports = permute;
